Replace the CanonicalMiscFilesProjectLoader
Rather than having two separate loaders one for canonical-project-
based projects and one for other projects, it's easier if we unify down
to one, and just return what the project should look like. This creates
a separate BuildHostProcessManager but the simplification there makes
that OK.
